requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
alembic>=1.12.0

# Frontend  
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.0.0
requests>=2.31.0
python-dateutil>=2.8.0

# Shared / Utils
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
